Fix gpt2, t5 and fnet under mixed precision
Also add a testing mode for mixed precision, though it does not fully
pass yet due to saving and other output errors.

+++ keras_nlp/layers/cached_multi_head_attention.py
@@ -84,7 +84,9 @@ class CachedMultiHeadAttention(keras.layers.MultiHeadAttention):
             value = dynamic_update_slice(value_cache, value, start)
             cache = tf.stack((key, value), axis=1)
 
-        query = tf.multiply(query, 1.0 / tf.math.sqrt(float(self._key_dim)))
+        query = tf.multiply(
+            query, 1.0 / tf.math.sqrt(tf.cast(self._key_dim, query.dtype))
+        )
         attention_scores = tf.einsum(self._dot_product_equation, key, query)
         attention_scores = self._masked_softmax(
             attention_scores, attention_mask
